Tailwind modals for select favorites and account
